+++ boards/components/src/lib.rs
@@ -12,6 +12,7 @@ pub mod console;
 pub mod crc;
 pub mod debug_queue;
 pub mod debug_writer;
+pub mod hd44780;
 pub mod isl29035;
 pub mod lldb;
 pub mod nrf51822;
