+++ capsules/src/lib.rs
@@ -26,12 +26,12 @@ pub mod fm25cl;
 pub mod fxos8700cq;
 pub mod gpio;
 pub mod gpio_async;
+pub mod hd44780;
 pub mod humidity;
 pub mod i2c_master;
 pub mod i2c_master_slave_driver;
 pub mod ieee802154;
 pub mod isl29035;
-pub mod lcd_1602;
 pub mod led;
 pub mod log;
 pub mod low_level_debug;
